requirements.txt
pywebview>=5.0
anthropic>=0.40
mss>=9.0
pywin32>=306
Pillow>=10.0
pynput>=1.7
keyring>=24.0
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
